feat: add sub collections

+++ apps/vendyx-api/src/business/collection/collection.service.ts
@@ -1,5 +1,5 @@
 import { Inject } from '@nestjs/common';
-import { PrismaPromise } from '@prisma/client';
+import { Collection, CollectionContentType, PrismaPromise } from '@prisma/client';
 
 import {
   CollectionFilters,
@@ -24,7 +24,8 @@ export class CollectionService {
       ...clean({ skip: input?.skip, take: input?.take }),
       where: {
         ...clean(input?.filters ?? {}),
-        name: input?.filters?.name ? { ...clean(input.filters.name), mode: 'insensitive' } : {}
+        name: input?.filters?.name ? { ...clean(input.filters.name), mode: 'insensitive' } : {},
+        parentId: null
       },
       orderBy: { createdAt: 'desc' }
     });
@@ -45,25 +46,55 @@ export class CollectionService {
   }
 
   async create(input: CreateCollectionInput) {
-    const { assets, products, ...rest } = input;
+    const { assets, products, contentType, subCollections, ...rest } = input;
     const slug = await this.validateAndParseSlug(input.name);
 
     return await this.prisma.collection.create({
       data: {
         ...clean(rest),
+        contentType,
         slug,
         assets: assets
           ? { create: assets.map(asset => ({ assetId: asset.id, order: 0 })) }
           : undefined,
-        products: products
-          ? { create: products.map(product => ({ productId: product, order: 0 })) }
-          : undefined
+        products:
+          products && contentType === CollectionContentType.PRODUCTS
+            ? { create: products.map(product => ({ productId: product, order: 0 })) }
+            : undefined,
+        subCollections:
+          subCollections && contentType === CollectionContentType.COLLECTIONS
+            ? { connect: subCollections.map(subCollection => ({ id: subCollection })) }
+            : undefined
       }
     });
   }
 
   async update(id: ID, input: UpdateCollectionInput) {
-    const { assets, products, ...rest } = input;
+    const { assets, products, subCollections, ...rest } = input;
+    const { contentType } = await this.prisma.collection.findUniqueOrThrow({ where: { id } });
+
+    const isModifyingSubCollection =
+      Array.isArray(subCollections) && contentType === CollectionContentType.COLLECTIONS;
+
+    let subCollectionStored: Collection[] = [];
+    let subCollectionsToStore: Collection[] = [];
+
+    // only fetch needed data for sub collections if we are modifying them
+    if (isModifyingSubCollection) {
+      subCollectionStored = await this.prisma.collection.findMany({
+        where: { parentId: id }
+      });
+      subCollectionsToStore = await this.prisma.collection.findMany({
+        where: { id: { in: subCollections ?? [] } }
+      });
+    }
+
+    const subCollectionsToCreate = subCollectionsToStore.filter(
+      s => s.contentType === CollectionContentType.PRODUCTS
+    );
+    const subCollectionsToRemove = subCollectionStored
+      .map(subCollection => subCollection.id)
+      .filter(subCollection => !subCollections?.includes(subCollection));
 
     const transaction: PrismaPromise<any>[] = [
       this.prisma.collection.update({
@@ -80,19 +111,25 @@ export class CollectionService {
                 }))
               }
             : undefined,
-          products: products
-            ? {
-                connectOrCreate: products.map(product => ({
-                  where: { productId_collectionId: { productId: product, collectionId: id } },
-                  create: { productId: product }
-                }))
-              }
-            : undefined
+          products:
+            products && contentType === CollectionContentType.PRODUCTS
+              ? {
+                  connectOrCreate: products.map(product => ({
+                    where: { productId_collectionId: { productId: product, collectionId: id } },
+                    create: { productId: product }
+                  }))
+                }
+              : undefined,
+          subCollections:
+            subCollectionsToCreate && contentType === CollectionContentType.COLLECTIONS
+              ? { connect: subCollectionsToCreate.map(subCollection => ({ id: subCollection.id })) }
+              : undefined
         }
       })
     ];
 
-    if (Array.isArray(products)) {
+    // if products is an array (meaning was modified) then remove all products that are not in the new list
+    if (Array.isArray(products) && contentType === CollectionContentType.PRODUCTS) {
       transaction.unshift(
         this.prisma.productCollection.deleteMany({
           where: { productId: { notIn: products ?? [] }, collectionId: id }
@@ -100,6 +137,16 @@ export class CollectionService {
       );
     }
 
+    if (subCollectionsToRemove.length && contentType === CollectionContentType.COLLECTIONS) {
+      transaction.unshift(
+        this.prisma.collection.updateMany({
+          where: { id: { in: subCollectionsToRemove } },
+          data: { parentId: null }
+        })
+      );
+    }
+
+    // if assets is an array (meaning was modified) then remove all assets that are not in the new list
     if (Array.isArray(assets)) {
       transaction.unshift(
         this.prisma.collectionAsset.deleteMany({
